+++ passkit/CHANGELOG.md
@@ -1,8 +1,13 @@
+## 0.0.5
+
+- Add signature validation
+- Rename `skipVerification` to `skipChecksumVerification`. This is a breaking change.
+
 ## 0.0.4
 
 - Add some missing `toJson` methods.
 - Improve readme
-- Add checksum verification (signature verification is still missing)
+- Add checksum validation (signature verification is still missing)
 - Change `webServiceUrl` from `String?` to `Uri?`. This is a breaking change
 - Fix building URLs for the PassKit web service
 - Remove `formatType` from `Barcode`. `Barcode.format` is an enum instead.